E-commerce-API-Project/ecomm_api/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.5
django-cors-headers==4.9.0
django-csp==4.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
iniconfig==2.1.0
mysql-connector-python==9.4.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
